move shanghai_block_zero_withdrawals test
